separate the express app for testing without listening to a port
